Rename run-flow.ts to create-and-run-flow.ts and update exports
