+++ packages/giselle-engine/src/core/flows/index.ts
@@ -1,4 +1,5 @@
 export * from "./configure-trigger";
+export * from "./create-and-run-flow";
 export * from "./create-run";
 export * from "./delete-trigger";
 export * from "./get-trigger";
@@ -6,5 +7,4 @@ export * from "./get-workspace-flow-runs";
 export * from "./patch-run";
 export * from "./resolve-trigger";
 export * from "./run/object";
-export * from "./run-flow";
 export * from "./set-trigger";
